Move common utility functions to a common library
