+++ .gitignore
@@ -13,7 +13,6 @@ __pycache__/
 /dist/
 /downloads/
 /eggs/
-/lib/
 /lib64/
 /parts/
 /sdist/
