+++ lib/ts_utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility functions for typeshed infrastructure scripts."""
