+++ tests/regr_test.py
@@ -21,8 +21,11 @@ from functools import partial
 from pathlib import Path
 from typing_extensions import TypeAlias
 
-from _metadata import get_recursive_requirements, read_metadata
-from _utils import (
+sys.path += ["lib"]
+
+
+from ts_utils.metadata import get_recursive_requirements, read_metadata  # noqa: E402
+from ts_utils.utils import (  # noqa: E402
     PYTHON_VERSION,
     TEST_CASES_DIR,
     DistributionTests,
