+++ tests/stubtest_third_party.py
@@ -13,8 +13,11 @@ from shutil import rmtree
 from textwrap import dedent
 from typing import NoReturn
 
-from _metadata import NoSuchStubError, get_recursive_requirements, read_metadata
-from _utils import (
+sys.path += ["lib"]
+
+
+from ts_utils.metadata import NoSuchStubError, get_recursive_requirements, read_metadata  # noqa: E402
+from ts_utils.utils import (  # noqa: E402
     PYTHON_VERSION,
     allowlist_stubtest_arguments,
     allowlists_path,
